+++ src/protocol/body/security_mode.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 bitflags! {
     #[derive(Serialize, Deserialize)]
-    pub struct SecurityMode: u8 {
+    pub struct SecurityMode: u16 {
         const NEGOTIATE_SIGNING_ENABLED = 0x01;
         const NEGOTIATE_SIGNING_REQUIRED = 0x02;
     }
